refactor: playerName, playerCards -> ParticipantName, Cards로 이름 변경
