+++ src/main/kotlin/blackjack/domain/Participant.kt
@@ -1,4 +1,11 @@
 package blackjack.domain
 
-class Participant {
+import blackjack.domain.card.Cards
+
+sealed class Participant(
+    private val state: State
+) {
+    fun cards(): Cards {
+        return state.cards()
+    }
 }
