+++ src/test/kotlin/blackjack/domain/player/PlayerTest.kt
@@ -1,6 +1,7 @@
 package blackjack.domain.player
 
 import blackjack.domain.ParticipantName
+import blackjack.domain.card.Cards
 import blackjack.helper.DeckHelper
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
@@ -10,7 +11,7 @@ class PlayerTest : BehaviorSpec({
     Given("덱이 주어졌을 때") {
         val player = Player(
             name = ParticipantName("플레이어1"),
-            cards = PlayerCards()
+            cards = Cards()
         )
         val deck = DeckHelper.createMockDeck()
         val card = deck.draw()
